Reverse InstructionBuffer after initial retrieval
When we grab a fresh copy of the list of instructions
to execute, we need to flip the buffer. This is so that
we can simply pop each instruction off the back which
proves to be cleaner than doing some shifting or
splicing of the buffer after retrieving the first element.