Introduce BusInterfaceUnit and rename modules
In the original 8086 architecture, the CPU had two
logical halves. The ExecutionUnit(EU) and the
BusInterfaceUnit(BIU). This patch renames the
Machine object to the more appropriate 'CPU'. The
ExecutionUnit has been shortened to EU and the
BusInterfaceUnit will be henceforth referenced as
'BIU'.

The BIU should handle all address translation and
memory accesses. The Segment registers and
Instruction Pointer exist within the logical BIU.
This means that the BIU handles instruction fetch
and queueing. Control flow will need to properly
notify the BIU of a change to the IP.